requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
tweepy==4.14.0
openai==1.3.5
python-telegram-bot==20.7
aiohttp==3.9.1
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6

